pyproject.toml
[tool.poetry]
name = "template-web-app-flask"
version = "0.1.0"
description = "A template for a web application built with Flask"
authors = ["Hassan Abedi <hassan.abedi.t@gmail.com>"]
maintainers = ["Hassan Abedi <hassan.abedi.t@gmail.com>"]
license = "MIT"
readme = "README.md"
include = ["README.md"]
packages = [{ include = "src", from = "." }]
repository = "https://github.com/habedi/template-web-app-flask"
documentation = "https://github.com/habedi/template-web-app-flask/blob/main/docs/README.md"
keywords = ["template", "python", "web", "flask"]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.1.0"
flask-migrate = "^4.1.0"
flask-wtf = "^1.2.2"
flask-admin = "^1.6.1"
flask-login = "^0.6.3"
sqlalchemy = "^2.0.39"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies] # Development dependencies (need `--with dev` flag)
poetry-dynamic-versioning = "^1.4.0"
pytest = "^8.0.1"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"
pytest-benchmark = "^5.1.0"
mypy = "^1.11.1"
ruff = "^0.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [".", "src"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
disallow_untyped_calls = true
strict_optional = true
warn_redundant_casts = true

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
versioning = "semver"  # Semantic Versioning

# Ruff configuration (Edit as needed)
[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    # Ignore the following additional directories and files
    "docs",
    "tests",
]
line-length = 100
indent-width = 4
src = ["src"]
target-version = "py310"

[tool.ruff.lint]
select = ["ANN", "D", "E", "F", "I"]
ignore = [
    # Ignore docstring errors
    "D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107",
]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = []
